requirements.txt
torch
torchaudio
espnet
google-generativeai
sounddevice
soundfile
pyaudio
transformers
librosa
numpy
scipy